+++ edison-jobs/src/main/java/de/otto/edison/jobs/eventbus/JobEvents.java
@@ -1,5 +1,14 @@
 package de.otto.edison.jobs.eventbus;
 
+import de.otto.edison.annotations.Beta;
+
+/**
+ * Delegates the calls of {@link JobEventPublisher#info(String)},
+ * {@link JobEventPublisher#warn(String)} and {@link JobEventPublisher#error(String)}
+ * to the correct JobEventPublisher of a job.
+ * This can be useful in methods outside the {@link de.otto.edison.jobs.service.JobRunnable#execute(JobEventPublisher)} method which do not know about the JobEventPublisher of a Job (e.g. stateless dependencies of a job).
+ */
+@Beta
 public class JobEvents {
 
     public static ThreadLocal<JobEventPublisher> jobEventPublisherThreadLocal = new InheritableThreadLocal<>();
@@ -38,7 +47,7 @@ public class JobEvents {
     }
 
     private static void checkInitialisation() {
-        if(jobEventPublisherThreadLocal.get() == null) {
+        if (jobEventPublisherThreadLocal.get() == null) {
             throw new IllegalStateException("JobEventPublisher has not been initialised. Try calling JobEvents.init() first");
         }
     }